Refactor LocalRoutingProvider for use by HotRod session management.
