+++ clustering/web/cache/src/main/java/org/wildfly/clustering/web/cache/routing/NullRouteLocator.java
@@ -20,24 +20,18 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.web.infinispan.routing;
+package org.wildfly.clustering.web.cache.routing;
 
 import org.wildfly.clustering.web.routing.RouteLocator;
 
 /**
- * Route locator that always returns the route of the local member.
+ * Route locator that always returns {@code null}.
  * @author Paul Ferraro
  */
-public class LocalRouteLocator implements RouteLocator {
-
-    private final String route;
-
-    public LocalRouteLocator(String route) {
-        this.route = route;
-    }
+public class NullRouteLocator implements RouteLocator {
 
     @Override
     public String locate(String sessionId) {
-        return this.route;
+        return null;
     }
 }
