+++ clustering/web/cache/src/test/java/org/wildfly/clustering/web/cache/routing/NullRouteLocatorTestCase.java
@@ -20,7 +20,7 @@
  * 02110-1301 USA, or see the FSF site: http://www.fsf.org.
  */
 
-package org.wildfly.clustering.web.infinispan.routing;
+package org.wildfly.clustering.web.cache.routing;
 
 import org.junit.Assert;
 import org.wildfly.clustering.web.routing.RouteLocator;
