+++ clustering/web/extension/src/main/java/org/wildfly/extension/clustering/web/deployment/DistributableWebDeploymentXMLReader.java
@@ -35,8 +35,8 @@ import org.jboss.as.server.deployment.DeploymentUnit;
 import org.jboss.as.server.deployment.jbossallxml.JBossAllXMLParser;
 import org.jboss.staxmapper.XMLElementReader;
 import org.jboss.staxmapper.XMLExtendedStreamReader;
-import org.wildfly.clustering.web.infinispan.routing.LocalRouteLocatorServiceConfiguratorFactory;
-import org.wildfly.clustering.web.infinispan.routing.NullRouteLocatorServiceConfiguratorFactory;
+import org.wildfly.clustering.web.cache.routing.LocalRouteLocatorServiceConfiguratorFactory;
+import org.wildfly.clustering.web.cache.routing.NullRouteLocatorServiceConfiguratorFactory;
 import org.wildfly.clustering.web.infinispan.routing.PrimaryOwnerRouteLocatorServiceConfiguratorFactory;
 import org.wildfly.clustering.web.infinispan.session.InfinispanSessionManagementConfiguration;
 import org.wildfly.clustering.web.infinispan.session.InfinispanSessionManagementProvider;
