+++ feature-pack/src/main/resources/modules/system/layers/base/org/wildfly/extension/clustering/web/main/module.xml
@@ -50,6 +50,7 @@
         <module name="org.wildfly.clustering.infinispan.client"/>
         <module name="org.wildfly.clustering.infinispan.spi"/>
         <module name="org.wildfly.clustering.service"/>
+        <module name="org.wildfly.clustering.web.cache"/>
         <module name="org.wildfly.clustering.web.infinispan"/>
         <module name="org.wildfly.clustering.web.spi"/>
     </dependencies>
